✨ feat(metric): add metric target exporter
