+++ pkg/service/service.go
@@ -101,7 +101,7 @@ func Run(ctx context.Context, opts *options.Options) error {
 
 	exporterHandler := exporter.NewHandler("gems_server", map[string]exporter.Collectorfunc{
 		"request":   exporter.NewRequestCollector(),
-		"alertrule": exporter.NewAlertRuleCollector(deps.Agentscli, deps.DyConfigProvider),
+		"alertrule": exporter.NewAlertAndMetricCollector(deps.Agentscli, deps.Databse, deps.DyConfigProvider),
 	})
 
 	// run
