+++ pkg/utils/agents/extend.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"net/http"
 	"net/url"
+	"sort"
 	"strings"
 	"time"
 
@@ -13,6 +14,7 @@ import (
 	"github.com/prometheus/alertmanager/pkg/labels"
 	alertmanagertypes "github.com/prometheus/alertmanager/types"
 	prommodel "github.com/prometheus/common/model"
+	"golang.org/x/sync/errgroup"
 	v1 "k8s.io/api/core/v1"
 	kerrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/types"
@@ -393,4 +395,31 @@ func (c *ExtendClient) CommitRawAlertResource(ctx context.Context, raw *promethe
 		return err
 	}
 	return c.Update(ctx, raw.AlertmanagerConfig)
+}
+
+func (c *ExtendClient) ListMetricTargets(ctx context.Context, namespace string) ([]*prometheus.MetricTarget, error) {
+	pms := monitoringv1.PodMonitorList{}
+	sms := monitoringv1.ServiceMonitorList{}
+	g := errgroup.Group{}
+	g.Go(func() error {
+		return c.List(ctx, &pms, client.InNamespace(namespace))
+	})
+	g.Go(func() error {
+		return c.List(ctx, &sms, client.InNamespace(namespace))
+	})
+	if err := g.Wait(); err != nil {
+		return nil, err
+	}
+
+	ret := []*prometheus.MetricTarget{}
+	for _, v := range pms.Items {
+		ret = append(ret, prometheus.ConvertToMetricTarget(v))
+	}
+	for _, v := range sms.Items {
+		ret = append(ret, prometheus.ConvertToMetricTarget(v))
+	}
+	sort.Slice(ret, func(i, j int) bool {
+		return strings.ToLower(ret[i].Name) < strings.ToLower(ret[j].Name)
+	})
+	return ret, nil
 }
